+++ src/player.h
@@ -21,8 +21,8 @@ class Player {
 
     double getX() const;
     double getY() const;
-    double getL() const;
-    double getC() const;
+    int getL() const;
+    int getC() const;
     double getGX() const;
     double getGY() const;
     int getID() const;
